+++ crate/server/src/database/tests/find_attributes_test.rs
@@ -56,6 +56,7 @@ pub async fn find_attributes<DB: Database>(
             Some(uid.clone()),
             owner,
             &symmetric_key,
+            symmetric_key.attributes()?,
             &HashSet::new(),
             db_params,
         )
@@ -97,6 +98,28 @@ pub async fn find_attributes<DB: Database>(
         .await?;
     assert_eq!(found.len(), 1);
     assert_eq!(found[0].0, uid);
+
+    // Define a link vector not present in any database objects
+    let link = vec![Link {
+        link_type: LinkType::ParentLink,
+        linked_object_identifier: LinkedObjectIdentifier::TextString("bar".to_string()),
+    }];
+
+    let researched_attributes = Some(Attributes {
+        object_type: Some(ObjectType::SymmetricKey),
+        link: Some(link.clone()),
+        ..Attributes::default()
+    });
+    let found = db
+        .find(
+            researched_attributes.as_ref(),
+            Some(StateEnumeration::Active),
+            owner,
+            true,
+            db_params,
+        )
+        .await?;
+    assert_eq!(found.len(), 0);
 
     Ok(())
 }